Check-in of messy script for generating Kconfig HTML
Please bear with me, this was written in a day without much thought.

The original plan was to make a tool that aids in comparing two
.configs, but at that time I did not know what Kconfiglib could offer.

This script parses the kernel configuration from Kconfig into a tree and
then generates nested HTML lists from it. Choices and menus can be
collapsed.

An ad-hoc helper bar on the right should help in navigating.

Could easily be modified to work with Python 2, but did not bother as it
is a write once hack.